% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{basic_bar}
\alias{basic_bar}
\title{Basic bar plot}
\usage{
basic_bar(data, title = "[Title]", xlab = "x")
}
\arguments{
\item{data}{The data frame}

\item{title}{The title of the plot}

\item{xlab}{The x-axis label. Defaults to \code{x}}
}
\value{
A ggplot object
}
\description{
Basic bar plot
}
