% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebooks.R
\name{build_codebook}
\alias{build_codebook}
\title{Build Quantitative Codebook}
\usage{
build_codebook(state, tool, outdir, quietly = TRUE)
}
\arguments{
\item{state}{The State where the Assessment was conducted}

\item{tool}{The specific category of tool that was administered}

\item{outdir}{The output directory}
}
\value{
The path of the rendered Rmarkdown file i.e. the return value for
\code{rmarkdown::render}. If building the document is interrupted by a
condition, the function fails silently and the condition object is returned.
}
\description{
Render an individual codebook for a given State and Tool
}
