% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dum.R
\name{build_dummy_flextable}
\alias{build_dummy_flextable}
\title{Make a dummy table}
\usage{
build_dummy_flextable(
  dic = NULL,
  rows = NULL,
  cols,
  top.header,
  multiresponse = T,
  caption = NULL,
  ...
)
}
\arguments{
\item{dic}{A \code{data.frame} containing the data dictionary}

\item{rows}{Numeric vector of the rows of interest in the dictionary}

\item{cols}{Character vector of columns that we want to use as focus}

\item{top.header}{Additional header row}

\item{multiresponse}{Boolean; whether table is for multi-response questions}

\item{caption}{The caption of the table}

\item{...}{Other parameters for internal use by the internal function
\code{make_table}}
}
\description{
Make a dummy table
}
