% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{build_report}
\alias{build_report}
\title{Build A Report}
\usage{
build_report(
  outdir,
  state,
  type = c("findings", "generic"),
  quietly = !interactive(),
  ...
)
}
\arguments{
\item{outdir}{Destination directory of generated report}

\item{state}{RAAMP State for which report is generated}

\item{type}{character - either \code{basic} or \code{findings}}

\item{quietly}{logical; should output be printed to the console?}

\item{...}{Other arguments passed to \code{rmarkdown::render}.}
}
\description{
Generate a report in MS Word format (\code{.docx} format) from an R Markdown
template.
}
\note{
By default, in interactive mode, the function will print progress
messages to the console, while it is quiet during a scripting session.
}
