% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_single_codebook}
\alias{build_single_codebook}
\title{Build codebook}
\usage{
build_single_codebook(state, tool, rmdfile, outputdir)
}
\arguments{
\item{state}{The State where the Assessment was conducted}

\item{tool}{The specific category of tool that was administered}

\item{rmdfile}{The template Rmarkdown file for building the codebook}

\item{outputdir}{The output directory}
}
\value{
The path of the rendered Rmarkdown file.
}
\description{
Render an individual codebook for a given State and Tool
}
