% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colCheck}
\alias{colCheck}
\title{Create Object Of Class colCheck}
\usage{
colCheck(regex, opts)
}
\arguments{
\item{regex}{A regular expression used for identifying the columns}

\item{opts}{TA list of character vectors, each of which are options for the
for the given question.}
}
\description{
Constructor for colCheck objects - a class that conveys information on
different multiple choice variables
}
