% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{collapse_multichoice}
\alias{collapse_multichoice}
\title{Collapse Columns for Multiiple-Choice Questions}
\usage{
collapse_multichoice(data, details)
}
\arguments{
\item{data}{An object of class \code{data.frame}}

\item{details}{An object of class \code{colCheck}}
}
\value{
An updated data frame
}
\description{
Collapse Columns for Multiiple-Choice Questions
}
\note{
All the newly created columns will be pushed to the right end of the
table.
}
