% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{collect_input}
\alias{collect_input}
\title{Collect User Input for Data Importation}
\usage{
collect_input()
}
\value{
A named list with 2 elements - one for each value collected.
}
\description{
Collects input from the user related to data for the
projects.
}
\details{
Two variables are collected from the user: the state of interest
and the kind of data i.e. services or capacity assessment. The input can
be collected interactively or passed as a command line argument.
}
\note{
When passing input via the shell or a script, \code{state} should
precede \code{filetype} in the argument list.
}
