% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{column_from_question}
\alias{column_from_question}
\title{Get Column from a question}
\usage{
column_from_question(data, ques)
}
\arguments{
\item{data}{The data frame}

\item{ques}{The question found in the variable label}
}
\description{
Get Column from a question
}
