% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_percent}
\alias{compute_percent}
\title{Convert a number to percentage}
\usage{
compute_percent(num)
}
\arguments{
\item{num}{The number to be converted}
}
\description{
Convert a number to percentage
}
