% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{confirm_state}
\alias{confirm_state}
\title{Get Project State(s)}
\usage{
confirm_state(x = NULL)
}
\arguments{
\item{x}{A string, usually partial match for a given State for that project.}
}
\value{
A character vector contain all project States when \code{x} is
\code{NULL} or a one of length \code{1L} based on the partial match.
}
\description{
Reels out the project states or a single one as selected by partial matching
}
\note{
This function is mostly used for conditional code chunk evaluation
in the R Markdown source files.
}
