% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{convert_word_to_text}
\alias{convert_word_to_text}
\title{Convert Word documents to a text format}
\usage{
convert_word_to_text(docpath, dest = ".", to_markdown = FALSE)
}
\arguments{
\item{docpath}{Path to an MS Word file}

\item{dest}{Destination folder}

\item{to_markdown}{Logical; whether to convert to markdown or not}
}
\value{
The path to the converted text file.
}
\description{
Convert Word documents to a text format
}
