% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_named_vector}
\alias{create_named_vector}
\title{creates a named vector from single string}
\usage{
create_named_vector(vec.list)
}
\arguments{
\item{vec.list}{A vector that contains a list of items for creating
a named vector}
}
\description{
creates a named vector from single string
}
