% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{create_referral_directory}
\alias{create_referral_directory}
\alias{create_capacity_assess}
\title{Create A Worksheet of the Referral Directory or Capacity Assessment}
\usage{
create_referral_directory(db, state, outdir, fname, ...)

create_capacity_assess(db, state, outdir, fname, ...)
}
\arguments{
\item{db}{The path to the database.}

\item{state}{The project State as a character string.}

\item{outdir}{The path to the directory where the output is to be saved.}

\item{fname}{An optional filename for the Excel file (is automatically
generated if not explicitly provided).}

\item{...}{Arguments passed on to \code{writeFormattedExcelSheet}.}
}
\value{
This function is used only for its side-effects.
}
\description{
Uses the project data to generate an MS Excel worksheet of the referral
directory for a givn project State.
}
