% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_value_label_pairs}
\alias{create_value_label_pairs}
\title{Creates value-label pairing for a given vector}
\usage{
create_value_label_pairs(variable, val.lab.obj)
}
\arguments{
\item{variable}{The data frame variable}

\item{val.lab.obj}{The object with the label:value pairs}
}
\description{
Creates value-label pairing for a given vector
}
