% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{custom_pkg_attach}
\alias{custom_pkg_attach}
\title{Customise Package Attachment}
\usage{
custom_pkg_attach(pkgs)
}
\arguments{
\item{pkgs}{A character vector with the names of the packages. An empty
string will yield an error}
}
\value{
None. This function is used for its side-effect only.
}
\description{
Provides a specific approach to attaching packages project-wide with a
focus on giving useful informative messages.
}
