% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{.removeSpaceForFilepath}
\alias{.removeSpaceForFilepath}
\title{Optimize Filenames Containing States}
\usage{
.removeSpaceForFilepath(state)
}
\arguments{
\item{state}{A character vector of length 1.}
}
\value{
The State with any speaces replaced with a hyphen
}
\description{
Remove space in a State's name for use with filepaths. This is to improve
portability for platforms and utilities e.g. \code{make} that do not work
well with filepaths that have spaces in them.
}
\details{
If \code{length(state) > 1L}, all additional elements are removed
silently. Also, \code{state} is expected to be a State in Nigeria; illegal
input will produce an error and the function will fail.
}
\examples{
  .removeSpaceForFilepath("Akwa Ibom")


}
