% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{dual_multiopts}
\alias{dual_multiopts}
\title{Produces dual outputs (table & plot) for multi-response variables}
\usage{
dual_multiopts(
  dat,
  indices,
  ...,
  size = getOption("jgbv.axis.text.size"),
  na.rm = TRUE
)
}
\arguments{
\item{dat}{The data frame}

\item{indices}{Columns to be used. Could be an integer vector (indices of
the columns) or a character vector of length 1 - a regular expression for
the columns to be selected.}

\item{...}{Arguments passed for internal use.}

\item{size}{The font size for the axis labels.}

\item{na.rm}{Logical; whether to keep NAs or not.}
}
\description{
Produces dual outputs (table & plot) for multi-response variables
}
