% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{dual_singleopts}
\alias{dual_singleopts}
\title{Makes outputs like above but for single response questions}
\usage{
dual_singleopts(dat, x, y = NULL, ..., na.rm = TRUE, use.table = FALSE)
}
\arguments{
\item{dat}{The data frame}

\item{x, y}{An integer or character vector of length \code{1L} for selecting
a column from \code{.data}.}

\item{...}{Arguments passed on to \code{applyFilter}.}

\item{na.rm}{Logical}

\item{use.table}{Logical. Whether a table object is created internally or
not for the handling of bivariate data.}
}
\description{
Makes outputs like above but for single response questions
}
