% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extend_single_listcol}
\alias{extend_single_listcol}
\title{Unnest a List Column}
\usage{
extend_single_listcol(data, var, chr)
}
\arguments{
\item{data}{The data frame}

\item{var}{The variable to be expanded}

\item{chr}{A character vector, which is the label of the value and which
will also be used to develop a new name for the column}
}
\value{
A modified data frame
}
\description{
'Open up list columns containing multiple choice responses by creating
a column for each option
}
