% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract}
\alias{extract}
\alias{extract_code}
\alias{extract_vector}
\title{Extracting code from REDCap Script  File}
\usage{
extract_code(rgx, txt)

extract_vector(str)
}
\arguments{
\item{rgx}{A reguar expression}

\item{txt}{A portion of text, supplied as a character vector}

\item{str}{A character string containing the code}
}
\value{
A character vector with the extracted code

A vector from the evaluated code
}
\description{
Extracting code from REDCap Script  File
}
\note{
This is an internal function but is exported to enable use in
the main RAAMP_GBV project.
}
