% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_code}
\alias{extract_code}
\title{Extract Code}
\usage{
extract_code(rgx, txt)
}
\arguments{
\item{rgx}{A reguar expression}

\item{txt}{A portion of text, supplied as a character vector}
}
\value{
A character vector with the extracted code
}
\description{
Extract Code
}
