% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_vector}
\alias{extract_vector}
\title{Extract Vector Code}
\usage{
extract_vector(str)
}
\arguments{
\item{str}{A string containing the code}
}
\value{
A vector from the evaluated code
}
\description{
Extract Vector Code
}
