% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{extract_zipfile}
\alias{extract_zipfile}
\title{Extract ZIP file}
\usage{
extract_zipfile(zipfile, state = NULL)
}
\arguments{
\item{zipfile}{Character vector of the ZIP file name}

\item{state}{State whose data is being extracted}
}
\value{
A data frame from  argument
set to \code{TRUE}
}
\description{
Extract a zip archive for a given state
}
