% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fetch_all_data}
\alias{fetch_all_data}
\title{Fetch All The Data}
\usage{
fetch_all_data(state, dir, sectors)
}
\arguments{
\item{state}{The RAAMP Assessment State e.g. Ogun}

\item{dir}{The directory where the data are saved.}

\item{sectors}{The sectors for which we are retrieving data}
}
\description{
Fetch All The Data
}
