% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{fetch_all_filepaths_named}
\alias{fetch_all_filepaths_named}
\title{Retrieve all files of a given type from a directory}
\usage{
fetch_all_filepaths_named(dir, file.ext = "R")
}
\arguments{
\item{dir}{character vector of directory path}

\item{file.ext}{The file extension; defaults to \emph{R}}
}
\value{
A character vector whose elements are the full paths of the
found files. If none is found, \code{character(0)}.
}
\description{
Lists files in a directory whose names have a certain file extension.
}
\note{
For the RAAMP-GBV RStudio project, the root directory is set to
the project directory.
}
