% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fetch_only_word_files}
\alias{fetch_only_word_files}
\title{Retrieve Word Files (containing transcripts)}
\usage{
fetch_only_word_files(dir)
}
\arguments{
\item{dir}{A diretory path}
}
\description{
Retrieve Word Files (containing transcripts)
}
