% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capneed-funs.R
\name{filterAndSelect}
\alias{filterAndSelect}
\title{Filter the Data and Select Only Columns Relevant to Capacity Assessment}
\usage{
filterAndSelect(df, service)
}
\arguments{
\item{df}{The data}

\item{service}{A string representing the service being filtered.}
}
\description{
Filter the Data and Select Only Columns Relevant to Capacity Assessment
}
