% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{filterData}
\alias{filterData}
\title{Filter Service Data}
\usage{
filterData(d, group.regex, srvtype)
}
\arguments{
\item{d}{The data frame}

\item{group.regex}{A regular expression that exactly matches the
prefix used for all the service categories.}

\item{srvtype}{The service type; also an exact match}
}
\value{
The modified data frame
}
\description{
Focus the data on a given GBV service category or sector
}
\note{
The function will preserve variable labels created with
\code{labelled::var_label}.
}
