% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_ques_with_var}
\alias{find_ques_with_var}
\title{Get a Question from the Variable name}
\usage{
find_ques_with_var(data, varname)
}
\arguments{
\item{data}{The data frame}

\item{varname}{The variable name}
}
\description{
Get a Question from the Variable name
}
