% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_var_across_datasets}
\alias{find_var_across_datasets}
\title{Check A Question Across Datasets}
\usage{
find_var_across_datasets(data.list, ques)
}
\arguments{
\item{data.list}{A list of data frames}

\item{ques}{The label containing the actual question}
}
\value{
A list of variables from each dataset containing the question.
}
\description{
Uses a question, held in the label to identify variables across different
data frames.
}
