% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_var_with_ques}
\alias{find_var_with_ques}
\title{Get a variable with it's question}
\usage{
find_var_with_ques(data, ques)
}
\arguments{
\item{data}{The data frame}

\item{ques}{The question found within the variable label}
}
\description{
Get a variable with it's question
}
