% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{fix_factors}
\alias{fix_factors}
\title{Cleanup Categorical Variables}
\usage{
fix_factors(data, newvars)
}
\arguments{
\item{data}{The data frame}

\item{newvars}{A named vector with the variable names as provided
by \code{options("jgbv.new.varnames")}.}
}
\value{
The modified data frame (if applicable).
}
\description{
Converts relevant character vectors in the dataset into factors.
}
