% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capneed-funs.R
\name{fixOffset}
\alias{fixOffset}
\alias{fixOffset.numeric}
\alias{fixOffset.character}
\alias{fixOffset.default}
\title{Fix offsets for some States}
\usage{
fixOffset(x, ...)

\method{fixOffset}{numeric}(x, df)

\method{fixOffset}{character}(x, df)

\method{fixOffset}{default}(x)
}
\arguments{
\item{x}{A vector}

\item{...}{Other arguments for methods}

\item{df}{A data frame}
}
\description{
Fix offsets for some States
}
