% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capneed-funs.R
\name{flex_datatable}
\alias{flex_datatable}
\title{Make a Table Specific to the Capacity Assessment Report}
\usage{
flex_datatable(data)
}
\arguments{
\item{data}{A data frame with the required data}
}
\description{
Make a Table Specific to the Capacity Assessment Report
}
