% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_all_codebooks}
\alias{generate_all_codebooks}
\title{Generate codebooks}
\usage{
generate_all_codebooks(states, sectors)
}
\arguments{
\item{states}{A character vectors with names of the Nigerian states of
interest as defined by the \code{states} object.}

\item{sectors}{A character vector with the labels of the various tools
as defined by the \code{tools.sector} object.}
}
\value{
A named list, with each element of the list being a character vector
of the filepaths of individual codebooks generated for a given State.
}
\description{
Automated generation of ALL the quantitative codebooks
}
