% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{generate_filename}
\alias{generate_filename}
\title{Generate File Names for Excel Outputs}
\usage{
generate_filename(state, type)
}
\arguments{
\item{state}{The State of interest.}

\item{type}{The type of output one of \code{refdir} (for Referral Directory)
or \code{capneed} (for Capacity Needs Assessment).}
}
\description{
Generate File Names for Excel Outputs
}
