% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqda.R
\name{get_codecat_dfs}
\alias{get_codecat_dfs}
\title{Dataframes for specific Code Categories}
\usage{
get_codecat_dfs(dir = here::here("data/qual/rqda"))
}
\arguments{
\item{dir}{A path to the directory housing the individual RQDA projects}
}
\value{
A list of data frames, one for each category
}
\description{
Gives a list of data frames, each for a given code category
}
