% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqda.R
\name{get_codings_master_df}
\alias{get_codings_master_df}
\title{Data related to codings}
\usage{
get_codings_master_df(projects)
}
\arguments{
\item{projects}{A character vector of the paths of one or more related
RQDA project databases}
}
\value{
A data frame containing the data from the combined projects
}
\description{
GEts a data frame of all the codings across different RQDA projects
}
