% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqda.R
\name{get_quotes}
\alias{get_quotes}
\title{Get Selected Codings}
\usage{
get_quotes(codedata, proj, code)
}
\arguments{
\item{codedata}{A data frame containing the codings, usually as an output
of \code{RQDA::getCodingsTable}.}

\item{proj}{An RQDA project.}

\item{code}{The code for which the codings are to be retrieved}
}
\value{
No value is returned but a window with the codings is opened
as a side effect.
}
\description{
Fetches the portions of text that were selected i.e. the codings for a
given code in an RQDA project
}
\note{
For this function to work properly, the RQDA GUI needs to have been
opened prior to its being called.
}
