% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqda.R
\name{get_rqda_projs}
\alias{get_rqda_projs}
\title{Get the paths to RQDA project}
\usage{
get_rqda_projs(datafolder = here::here("data/qual/rqda"))
}
\arguments{
\item{datafolder}{The directory housing the projects}
}
\description{
Collect RQDA projects developed by different collaborators
}
