% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_service_data}
\alias{get_service_data}
\title{Get Data on Services Offered}
\usage{
get_service_data(data, ques, options)
}
\arguments{
\item{data}{A data frame}

\item{ques}{The question (found in the variable label itself)}

\item{options}{The service options for that given question and tool}
}
\description{
Get Data on Services Offered
}
