% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{import_redcap_csv}
\alias{import_redcap_csv}
\title{Imports CSV downloaded from REDCap into R session}
\usage{
import_redcap_csv(path.list, pattern)
}
\arguments{
\item{path.list}{A character vector of filepaths}

\item{pattern}{A regular expression for searching for a given CSV file}
}
\value{
A dataframe; \code{stringsAsFactors} is set to \code{FALSE}
}
\description{
Imports CSV downloaded from REDCap into R session
}
