% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R, R/transform.R
\name{input_state}
\alias{input_state}
\title{collect Name of RAAMP State}
\usage{
input_state()

input_state()
}
\description{
Collect the name of a given State for the purpose of using it to build a
given part of the project that is state-dependent.

Collect the name of a given State for the purpose of using it to build a
given part of the project that is state-dependent.
}
\details{
This function will collect input both from the shell and during
interactive R sessions. Also, if a State's name is supplied as an argument
to \code{Rscript}, it will read it and apply it as required.

This function will collect input both from the shell and during
interactive R sessions. Also, if a State's name is supplied as an argument
to \code{Rscript}, it will read it and apply it as required.
}
