% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interview.R
\name{interview}
\alias{interview}
\title{Interviews}
\usage{
interview(questions)
}
\arguments{
\item{questions}{A character vector of the question(s)}
}
\value{
The function returns an object of class \code{interview}.
}
\description{
This function constructs an S3 object called \code{interview}, which
represents the questions asked in the tools. The object also has
regular expressions, one for each questions, that can be utilized in
carrying out string matching during the analysis.
}
\examples{
obj <- interview("What are the issue(s)?")
obj$question
obj$regex

}
