% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_project_state}
\alias{is_project_state}
\title{Test For RAAMP-GBV Project State}
\usage{
is_project_state(str)
}
\arguments{
\item{str}{A character vector of length 1.}
}
\value{
A boolean value \code{TRUE} or \code{FALSE}
}
\description{
Checks whether a string represents a RAAMP-GBV project state
}
