% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{raampStates}
\alias{raampStates}
\alias{tool.sectors}
\title{Builtin Objects For JHPIEGO GBV Projects}
\format{
An object of class \code{character} of length 4.

An object of class \code{Tools} of length 7.
}
\usage{
raampStates

tool.sectors
}
\description{
Builtin Objects For JHPIEGO GBV Projects
}
\keyword{datasets}
