% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Launch the Shiny App}
\usage{
launch_shiny(...)
}
\arguments{
\item{...}{Arguments (other than \code{appDir}) passed to
\code{link[shiny::runApp]{name}}}
}
\description{
Lauches the Shiny application in the session, or if not interactive,
in the system's default browser
}
