% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_files_pattern}
\alias{list_files_pattern}
\title{Conveniently list files in a directory}
\usage{
list_files_pattern(dir = ".", pattern)
}
\arguments{
\item{dir}{The path of the directory}

\item{pattern}{A regular expression for selecting files}
}
\value{
A character vector with all the files whose paths match the
\code{pattern}. Otherwise returns \code{character(0)}.
}
\description{
Conveniently list files in a directory
}
