% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_data}
\alias{load_data}
\title{Loads data from the project database(s)}
\usage{
load_data(path, state, type = c("services", "capacity"))
}
\arguments{
\item{path}{The path to the data(base) file.}

\item{state}{The State for which the data are retrieved.}

\item{type}{The kind of data being retrieved (see \italics{Details})}

\item{...}{Arguments for methods}
}
\description{
Loads data from the project database(s)
}
\details{
There are two possible values for the \code{type} argument:
\code{services} for retrieving data on GBV services mapping and
\code{capacity} for data on capacity assessment.
}
