% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{load_shiny_data}
\alias{load_shiny_data}
\title{Load data for Shiny app}
\usage{
load_shiny_data(state)
}
\arguments{
\item{state}{The RAAMP-GBV Assessment State for which data are to be
visualised}
}
\value{
A data frame.
}
\description{
Loads data that are specifically tailored for the project's Shiny application.
}
