% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{make_dir_tree}
\alias{make_dir_tree}
\title{Make a Representation of the Project's Standard Directory Tree}
\usage{
make_dir_tree()
}
\description{
Make a Representation of the Project's Standard Directory Tree
}
