% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dum.R
\name{make_just_yesno}
\alias{make_just_yesno}
\title{Make a Yes/No Dummy Table}
\usage{
make_just_yesno(columns, all.lgas)
}
\arguments{
\item{columns}{The colums to be used}

\item{all.lgas}{The LGAs in the State}
}
\description{
Make a Yes/No Dummy Table
}
