% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_kable}
\alias{make_kable}
\title{Make a kable table}
\usage{
make_kable(row, column, opt = NULL)
}
\arguments{
\item{row}{A vector for the rows.}

\item{column}{A character vector of the column names}

\item{label}{A label to be applied to the table; as in \code{knitr::kable}}

\item{caption}{A caption. Used as in \emph{label} above.}
}
\description{
Create a table formatted for markdown
}
