% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dum.R
\name{make_multiresp}
\alias{make_multiresp}
\title{Make a multi-reponse dummy table using established headers}
\usage{
make_multiresp(hdrs)
}
\arguments{
\item{hdrs}{Character vector with the names of the labels for each response}
}
\value{
A data frame with summary of multiple responses
}
\description{
Make a multi-reponse dummy table using established headers
}
