% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capneed-funs.R
\name{makeTable}
\alias{makeTable}
\title{Make a table of Capacity (Training) Needs}
\usage{
makeTable(dt, caption = NULL)
}
\arguments{
\item{dt}{A data frame for the type of service provider}

\item{caption}{A caption for the table}
}
\description{
Make a table of Capacity (Training) Needs
}
\note{
This function depends on a global variable \code{allcap}, which is
found in the workspace of the project and activated by the appropropriate
script. Hence it is for now unfortunately tied to is.
}
