% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{manually_check_filenames}
\alias{manually_check_filenames}
\title{Manually Check File Names}
\usage{
manually_check_filenames(directory)
}
\arguments{
\item{directory}{A directory path}
}
\description{
Manually Check File Names
}
