% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{modify_varnames}
\alias{modify_varnames}
\title{Change Variable Names}
\usage{
modify_varnames(as_list = FALSE)
}
\arguments{
\item{as_list}{Whether to present the results as a list. Otherwise as an
atomic vector}
}
\description{
Modifies the variable names and returns a list with both the old and new
}
