% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{multiresponse_by_lga}
\alias{multiresponse_by_lga}
\title{Make Multiple Response Tabulation Disaggregated By Local Government Area}
\usage{
multiresponse_by_lga(data, indices, text.removed = NULL)
}
\arguments{
\item{data}{The data frame}

\item{indices}{A numeric vector with the indices of the columns with
the requested variable.}

\item{text.removed}{A string. Portion of text that is to be removed from
the variable elements (optional).}
}
\value{
An object of class \code{data.frame}.
}
\description{
Make Multiple Response Tabulation Disaggregated By Local Government Area
}
