% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{my_autonum}
\alias{my_autonum}
\title{Generate Auto-numbers for Tables}
\usage{
my_autonum(bookmark = NULL, ...)
}
\arguments{
\item{bookmark}{A string to represent the bookmark used in a given document}

\item{...}{Other arguments passed on to \link{\code{[officer]{run_autonum}}}.}
}
\value{
See the documentation for \code{\link[officer]{run_autonum}}.
}
\description{
Generate Auto-numbers for Tables
}
