% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{not_multichoice}
\alias{not_multichoice}
\title{Check which columns are multiple choice}
\usage{
not_multichoice(x)
}
\arguments{
\item{x}{A column, represented as a vector}
}
\description{
Check which columns are multiple choice
}
