% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqda.R
\name{obtain_rqda_codes}
\alias{obtain_rqda_codes}
\title{Get the coding table from RQDA projects from collaborators}
\usage{
obtain_rqda_codes(path)
}
\arguments{
\item{path}{A path to a directory, zip file or RQDA projecct file}
}
\value{
The coding table as an R \code{data.frame}.
}
\description{
This function is designed to just fetch RQDA projects
with as little hassle as possible. So whether it's a directory, or
a compressed archive or an RQDA project file itself, the codes will
be extracted.

In the event that there are several RQDA files in the
accessed directory, they will all be extracted and merged into one
single table. No further transformation is carried out, thereafter.
}
