% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{odd_even_seq}
\alias{odd_even_seq}
\title{Creates sequences of odd or even value}
\usage{
odd_even_seq(vec, type = c("odd", "even"))
}
\arguments{
\item{vec}{The vector to be checked}

\item{type}{Whether to follow an odd or even sequence}
}
\description{
Creates sequences of odd or even value
}
