% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{show_output.default}
\alias{show_output.default}
\title{Creates an expression from x, the function for building tables,
and draws plot as well as produces a matching object
of class 'flextable'}
\usage{
\method{show_output}{default}(x, type = c("multiplot", "yesno"), ...)
}
\arguments{
\item{x}{An expression that produces the table for the relevant data}

\item{type}{The type of output. Options are \code{multiplot} and
\code{yesno}}

\item{...}{Arguments passed to \code{plot_multiplot} or \code{plot_yesno}}
}
\description{
Creates an expression from x, the function for building tables,
and draws plot as well as produces a matching object
of class 'flextable'
}
