% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commence.R
\name{outputs_commencement}
\alias{outputs_commencement}
\title{Ouputs on Opening and Commencement of GBV Services for a State}
\usage{
outputs_commencement(data, open, gbv, format = "\%Y")
}
\arguments{
\item{open}{The column with the opening data (an integer).}

\item{gbv}{The column with the date when GBV service was commenced
(an integer).}

\item{format}{The format to be used in extracting the date as in
\code{base::format}.}

\item{.data}{The data frame}
}
\value{
A list with 2 elements: An object of class \code{flextable} and
another of class \code{ggplot}, respectively called \code{table} and
\code{plot}.
}
\description{
Ouputs on Opening and Commencement of GBV Services for a State
}
