% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{pick_one_state}
\alias{pick_one_state}
\title{Interactively Select a State}
\usage{
pick_one_state()
}
\description{
Uses global options to identify the States for a given project and prompts
the user to select one.
}
