% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pickout_cols_and_transform}
\alias{pickout_cols_and_transform}
\title{Pick out a Colum and Transform it}
\usage{
pickout_cols_and_transform(data, chklist)
}
\description{
Pick out a Colum and Transform it
}
