% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{plot_services_by_lga}
\alias{plot_services_by_lga}
\title{Draw A Plot of Service Types By LGA}
\usage{
plot_services_by_lga(dat, labs = NULL)
}
\arguments{
\item{dat}{A data frame; most appropriately one generated by
\code{link{multiresponse_by_lga}}}

\item{labs}{The labels representing each service category}
}
\value{
An object of class \code{ggplot}.
}
\description{
Draw A Plot of Service Types By LGA
}
\details{
To use the labels correctly, one may first run the function with
the default values and see if there is a need to fine-tune.
}
