% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{prep_ref_directory}
\alias{prep_ref_directory}
\title{Make the Referral Directory}
\usage{
prep_ref_directory(df, state, indices, serv.cols, day.pattern, namelist)
}
\arguments{
\item{df}{The cleaned data from the State.}

\item{state}{The State under study.}

\item{indices}{A numeric vector of all the column indices that will be used to
create the tabulation.}

\item{serv.cols}{The columns that contain the interventions provided. Used either
as an atomic vector with indices, a regular expression or the actual column names.
by each category of service provider.}

\item{day.pattern}{A regular expression that represents the \code{colnames} of
all those that represent the day a facility operates.}

\item{namelist}{A named list with column names as follows:
\itemize{
  \item \strong{orgname} - Name of the organization.
  \item \strong{orgphone} - Phone number of the organization.
  \item \strong{gbvcontat} - Contact details of the GBV focal person.
}
In all instances, the value is \emph{the column name} for that variable.}
}
\value{
A modified data frame containing the information required of the
State's referral directory.
}
\description{
Creates a data frame of the referral directory of the State, which can be
used to create other more sylistically appealing tables
}
