% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{question_in_all_sectors}
\alias{question_in_all_sectors}
\title{Check whether a question is found across the sectors}
\usage{
question_in_all_sectors(quest, table)
}
\arguments{
\item{quest}{The question asked in the survey (acutally a regular expression)}

\item{table}{A list where each element is a character vector of all the
questions - extracted from the data frame labels.}
}
\value{
A named logical vector for all the sectors checked.
}
\description{
Check whether a question is found across the sectors
}
