% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{data}
\name{raampStates}
\alias{raampStates}
\alias{tool.sectors}
\title{Builtin Objects For RAAMP-GBV Project}
\format{An object of class \code{character} of length 4.}
\usage{
raampStates

tool.sectors
}
\description{
Builtin Objects For RAAMP-GBV Project
}
\keyword{datasets}
