% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_from_db}
\alias{read_from_db}
\title{Read Data From A Database}
\usage{
read_from_db(db, tbl, ...)
}
\arguments{
\item{db}{The path to the database file.}

\item{tbl}{The name of the table to be accessed.}

\item{...}{Additional arguments passed on to \code{DBI::dbConnect}.}
}
\description{
Reads data from the database and may carry out some minor transformations
in the process.
}
