% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{removed_variables}
\alias{removed_variables}
\title{Indices for columns that are removed from the variables}
\usage{
removed_variables(state = NULL)
}
\arguments{
\item{state}{The project state}
}
\value{
An integer vector of indices for columns to be dropped. For States
that are not applicable, it returns \code{NULL}.
}
\description{
A helper function for variable removal that may be necessary for
some of the data for specific States.
}
