% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{repair_lga_spelling}
\alias{repair_lga_spelling}
\title{Repair Poorly Spelt Nigeria Local Government Areas in the Project}
\usage{
repair_lga_spelling(x, state)
}
\arguments{
\item{x}{The character vector to be checked.}

\item{state}{The state whose LGAs will be checked.}
}
\description{
A wrapper for \code{naijR::fix_region} for dealing with misspelt
Local Government Areas.
}
