% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{replicate_rqdaproj}
\alias{replicate_rqdaproj}
\title{Creates a new RQDA project from the existing base project}
\usage{
replicate_rqdaproj(serialno, dir, state)
}
\arguments{
\item{serialno}{A serial number for the project.}

\item{dir}{The directory where the project is to be saved to.}

\item{state}{A state of Nigeria where the data were collected from.}
}
\value{
The value of \code{\link[base]{file.copy}}.
}
\description{
Creates a new RQDA project from the existing base project
}
