% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqda.R
\name{retrieve_codingtable}
\alias{retrieve_codingtable}
\alias{get_codings_master_df}
\title{Data related to codings}
\usage{
retrieve_codingtable(proj, query = character())

get_codings_master_df(proj)
}
\arguments{
\item{proj}{A character vector of the paths of one or more related
RQDA project databases}

\item{query}{A valid SQLite (i.e. including a terminating semicolon) in
the form of a string. Character vectors longer than \code{1L} are
truncated.}
}
\value{
A data frame containing the data, either singly or stacked
when more than one are combined.
}
\description{
Gets a data frame of all the codings across one or more RQDA projects.
}
\details{

}
