% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_file}
\alias{select_file}
\title{Select a file from a directory}
\usage{
select_file(path.list, pattern, type = c("R", "csv"))
}
\arguments{
\item{path.list}{A character vector of paths}

\item{pattern}{A regular expression}

\item{type}{The file format. The functions only support \emph{.R} and
\emph{CSV} files, with R scripts being the default format.}
}
\description{
Selects a file from a vector of file names based on a regex pattern
}
