% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{set_logical_with_label}
\alias{set_logical_with_label}
\title{Replaces the logical TRUE in the column with its label. This
features prominently in those columns that have to do with
the type of services/interventions carried out by facilities}
\usage{
set_logical_with_label(column)
}
\arguments{
\item{column}{A column of the data frame that inherits from class
\code{labelled}.}
}
\value{
The transformed column, now of type \code{character}.
}
\description{
Replaces the logical TRUE in the column with its label. This
features prominently in those columns that have to do with
the type of services/interventions carried out by facilities
}
