% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{SheetName}
\alias{SheetName}
\title{Gets or sets worksheet names}
\usage{
SheetName(type = c("capneeds", "refdir"))
}
\arguments{
\item{type}{Either \code{capneeds} or \code{refdir}. Partial matching is
allowed.}
}
\value{
A character vector of length 1L, nameing the sheet.
}
\description{
A convenience function for consistent naming of spreadsheets for the
relevant outputs.
}
