% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebooks.R
\name{show_codebook}
\alias{show_codebook}
\title{View a codebook}
\usage{
show_codebook(state, sector)
}
\arguments{
\item{state}{One of the project states}

\item{sector}{The sector of interest as defined by \code{tool.sectors}}
}
\description{
Opens up a quantitative codebook into the default web browser.
}
