% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{show_output}
\alias{show_output}
\title{Display output}
\usage{
show_output(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
Displays some output such as a plot/
}
