% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{single_sector_services_plot}
\alias{single_sector_services_plot}
\title{Plot the Services for a Single Sector Selected}
\usage{
single_sector_services_plot(data)
}
\arguments{
\item{data}{The data frame}
}
\description{
Plot the Services for a Single Sector Selected
}
