% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{spaceAndTitle}
\alias{spaceAndTitle}
\title{Modify Strings}
\usage{
spaceAndTitle(str)
}
\arguments{
\item{str}{A character vector of strings that will be modified.}
}
\value{
A character vector of the same length as \code{str}
}
\description{
Creates spaces for underscores and changes text to title case
}
