% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capneed-funs.R
\name{statGBV}
\alias{statGBV}
\title{Generate Inline Statisics for the Report}
\usage{
statGBV(df, stat = c("sum", "perc", "total"), var = character())
}
\arguments{
\item{df}{The data}

\item{stat}{The statistic to be used}

\item{var}{The variable of interest.}
}
\description{
Generate Inline Statisics for the Report
}
