% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{store_rqdacodes}
\alias{store_rqdacodes}
\title{Save New Codes To Disk}
\usage{
store_rqdacodes(codes, dir)
}
\arguments{
\item{codes}{An object of class \code{LocalCodes} which are created by the
function \code{add_rqdacodes}.}

\item{dir}{Path to the directory to which to save the persistent data.
Defaults to the folder containing the RQDA project file.}
}
\description{
Saves a local representation of the codes of the RQDA project for viewing
and easy sharing in version control
}
