% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_multiopt}
\alias{table_multiopt}
\alias{table_yesno}
\title{Make a frequency tabulation of multi-response questions}
\usage{
table_multiopt(
  x,
  dictionary = NULL,
  indices,
  use.regex = TRUE,
  data.only = FALSE,
  redcap = getOption("data.on.redcap"),
  ...
)

table_yesno(data, col, data.only = FALSE, ...)
}
\arguments{
\item{x}{The data}

\item{dictionary}{The data dictionary}

\item{indices}{Numeric vector representing indices of columns chosen}

\item{use.regex}{Logical; whether to use label patterns to determine
what the options should be for the responses}

\item{data.only}{Return only the \code{data.frame}?}

\item{...}{Additional response options, if they exist.}

\item{data}{The data.}

\item{col}{The column with the responses}
}
\description{
Make a frequency tabulation of multi-response questions

Make a frequency tabulation that for variables with Yes/No responses
}
\note{
To control what happens via the \code{use.regex} argument, the
project has to set this option, ideally in the local \code{.Rprofile} to
a desired value. For instance, to set a project-wide default value of
\code{FALSE} for this argument, set \code{options(use.regex = FALSE)}, and
that way, there is no need to set the argument, each time the function
is called.
}
