% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{transform_bool_to_logical}
\alias{transform_bool_to_logical}
\title{Change Binary Values to Logical}
\usage{
transform_bool_to_logical(data)
}
\arguments{
\item{data}{The data frame}
}
\value{
The data frame, now modified (if applicable).
}
\description{
Set the Columns that have 1/0 values to T/F instead, notably those
for multiple response variables
}
