% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_checked}
\alias{transform_checked}
\title{Transform data for "Checked/Unchecked"}
\usage{
transform_checked(data, vec)
}
\arguments{
\item{data}{The data frame}

\item{vec}{The given vector}
}
\description{
Transform data for "Checked/Unchecked"
}
