% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_sector_data}
\alias{transform_sector_data}
\title{Transform the data for a given sector}
\usage{
transform_sector_data(data, regexes, options)
}
\arguments{
\item{data}{The data frame with the data for a given sector}

\item{regexes}{Regular expression defining patterns for manipulating
columns for multiple choice questions}

\item{options}{The choices}
}
\description{
Transform the data for a given sector
}
