% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transform_to_factor}
\alias{transform_to_factor}
\title{Selectively transform a variable into a factor}
\usage{
transform_to_factor(x)
}
\arguments{
\item{x}{An atomic vector, usually part of a data frame}
}
\description{
Selectively transform a variable into a factor
}
