% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{view_rqdacodes}
\alias{view_rqdacodes}
\title{Display current RQDA codes for the project}
\usage{
view_rqdacodes(dir)
}
\arguments{
\item{dir}{Path to the directory to which to save the persistent data.
Defaults to the folder containing the RQDA project file.}
}
\description{
Display current RQDA codes for the project
}
