% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{write_files_to_project}
\alias{write_files_to_project}
\title{Write Text Files to an RQDA Project}
\usage{
write_files_to_project(proj, pattern = NULL, sourcedir = NULL)
}
\arguments{
\item{proj}{The path to the RQDA project.}

\item{pattern}{A regular expression for matching selected files.}

\item{sourcedir}{The path to the directory containing the text files.}

\item{ignore_case}{Logical; whether to consider the case when matching.}
}
\value{
The function has the side-effect of altering the state of the
indicated RQDA project.
}
\description{
Selects a number of files to RQDA based on a regular expression
matched on the filename.
}
\note{
The files to be imported are expected to be in plain text format.
}
