% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{writeFormattedExcelSheet}
\alias{writeFormattedExcelSheet}
\title{Write an Excel Sheet for Project Outputs}
\usage{
writeFormattedExcelSheet(
  data,
  path,
  sheet,
  header.fill,
  header.font.colour,
  na.string = ""
)
}
\arguments{
\item{data}{The data frame made with preparatory functions}

\item{path}{The path of the Excel file to be created.}

\item{sheet}{The name of the worksheet.}

\item{header.fill}{The colour for the header.}

\item{header.font.colour}{The colour of the header text.}

\item{na.string}{How to represent missing values}
}
\value{
No return value. Used for its side effects.
}
\description{
This function actually finalises the process of getting an Excel
worksheet ready for the Referral Direcories and Capacity Needs
Assessment tables.
}
