% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_rqdacodes}
\alias{add_rqdacodes}
\title{Add some codes to the RQDA project}
\usage{
add_rqdacodes(codes)
}
\arguments{
\item{codes}{A character vector of the new codes that are to be added.}
}
\description{
Add some codes to the RQDA project
}
\note{
The rationale behind creating this function is the fact that merging
the binary RQDA in Git version control is not feasible. Thus, codes can be
added to the existing ones and represented in the R session for further
uses downstream.
}
