% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{dual_multiopts}
\alias{dual_multiopts}
\title{Produces dual outputs (table & plot) for multi-response variables}
\usage{
dual_multiopts(dat, indices, ..., na.rm = TRUE)
}
\arguments{
\item{dat}{The data frame}

\item{indices}{Columns to be used}

\item{...}{Arguments passed for internal used}

\item{na.rm}{Logical; whether to keep NAs or not.}
}
\description{
Produces dual outputs (table & plot) for multi-response variables
}
