% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_var_labels}
\alias{get_var_labels}
\title{Get Variable Labels}
\usage{
get_var_labels(data, ind)
}
\arguments{
\item{data}{A data frame}

\item{ind}{A numeric or character vector representing columns}
}
\value{
A character vector of label names
}
\description{
Retrieve all the labels of designated columns of a data frame
}
