% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{import_data}
\alias{import_data}
\title{Import Raw Data Into the Project}
\usage{
import_data(db, modlist, state, filetype)
}
\arguments{
\item{db}{Path to an SQLite database. If non-existent, will attempt to
create one.}

\item{modlist}{A list of objects of class \code{VarModifier}.}

\item{state}{The project state.}

\item{filetype}{The kind of data being imported. Should be one of
\emph{Capacity} or \emph{Services} for data on capacity assessment and for
GBV service mapping, respectively.}
}
\value{
After saving the data table, return as a data frame.
}
\description{
Import Raw Data Into the Project
}
