% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{import_data}
\alias{import_data}
\alias{read_in_excel_data}
\title{Import Raw Data Into the Project}
\usage{
import_data(dir, db = NULL, modlist = NULL, state, filetype, ..., na.strings)

read_in_excel_data(
  dir,
  state,
  filetype,
  drop.c = NULL,
  drop.v = NULL,
  nvars = NULL,
  na.strings = ""
)
}
\arguments{
\item{dir}{A directory used for the downloads, which usually has
subdirectories - one for each project State.}

\item{db}{Path to an SQLite database. If non-existent, will attempt to
create one. If \code{NULL}, the data are returned without being saved.}

\item{modlist}{An optional list of objects of class \code{VarModifier}.}

\item{state}{The project state.}

\item{filetype}{The kind of data being imported. Should be one of
\emph{Capacity} or \emph{Services} for data on capacity assessment and for
GBV service mapping, respectively.}

\item{...}{Arguments passed to \code{read_in_excel_data}.}

\item{na.strings}{Strings in the Excel sheet to be interpreted as \code{MA}.}

\item{drop.c, drop.v}{Character or integer vector of columns or variables to
be dropped from the data table or list of variable names, respectively.}

\item{nvars}{The number of columns in the final data frame.}
}
\value{
Both functions return a \code{data.frame} with labelled variables.
\code{import_data} does so invisibly.
}
\description{
Import Raw Data Into the Project
}
\note{
\code{import_data} wraps \code{read_in_excel_data}, as it carries
out additional processing of the data read from the spreadsheets.
\code{nvars} is usually set for the capacity assessment data, which tend
to have superfluous columns.
}
