% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_transcripts}
\alias{import_transcripts}
\alias{lga_matching_function}
\title{Extract interview transcripts for RQDA}
\usage{
import_transcripts(
  zipfile = NULL,
  patterns = NULL,
  ignore_case = FALSE,
  no. = NULL
)

lga_matching_function(state, ...)
}
\arguments{
\item{zipfile}{Path to a ZIP archive containing interview transcripts. When
\code{NULL}, a file selection dialog is presented to the user - this only works
in interactive sessions.}

\item{patterns}{A regular expression or matching string for selecting LGAs.}

\item{ignore_case}{Logical; whether to consider the case when matching.}

\item{no.}{The number of projects to be created as a numeric vector of length 1.}

\item{state}{The project State for which the matches/patters are to be
extracted.}

\item{...}{Optional list of States to be manually provided for pattern
selection.}
}
\value{
For \code{lga_match_function}, a function that takes a list of patterns
for different states. No value is returned by \code{import_transcriipts}


}
\description{
A script to extract the zip archive containing interview transcripts
}
\details{
First we pick the archive interactively and unzip it into the same
directory containing the archive. After establishing the directory
for the extraction, we modify the names of the files by removing
all spaces therein. Then we move them into the State-specific
folder. When finalising, we do some checks that everything went well
and we remove the directory that was created for the extraction.
}
\note{
The list of project States is set internally by
\code{getOption("jgbv.project.states")}. This option is used by the
GBV assessment project beginning from 2021, and does not apply to earlier
ones.

To get the appropriate patterns or matches to be used for the extracting
function, there is need to inspect the list of transcript files. There, it
will be discovered how they are named and differentiated by LGA.
}
