% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_data}
\alias{load_data}
\alias{save_table}
\title{Save/Load Data To Or From The Project Database(s)}
\usage{
load_data(
  path,
  state,
  type = c("services", "capacity"),
  vars = getOption("jgbv.new.varnames")
)

save_table(df, state, type = c("services", "capacity"), path)
}
\arguments{
\item{path}{The path to the data(base) file.}

\item{state}{The State for which the data are retrieved.}

\item{type}{The kind of data being retrieved (see \emph{Details})}

\item{vars}{A vector of variable names.Defaults to names provided via
\code{options}.}

\item{df}{A project data frame that is to be saved.}
}
\value{
For \code{load_data},aA data frame with individual columns of
class "labelled".
}
\description{
Save/Load Data To Or From The Project Database(s)
}
\details{
There are two possible values for the \code{type} argument:
\code{services} for retrieving data on GBV services mapping and
\code{capacity} for data on capacity assessment.
}
