% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_date}
\alias{make_date}
\title{Manipulate A Vector For Getting Correct Date-Time Types}
\usage{
make_date(x)
}
\arguments{
\item{x}{The vector to be modified.}
}
\value{
The modified vector usually of class \code{Date} or \code{POSIXct}.
}
\description{
Manipulate A Vector For Getting Correct Date-Time Types
}
