% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dum.R
\name{make_kable}
\alias{make_kable}
\title{Make a kable table}
\usage{
make_kable(row, column, opt = NULL)
}
\arguments{
\item{row}{A vector for the rows.}

\item{column}{A character vector of the column names}

\item{opt}{The options for the dummy table}
}
\description{
Create a table formatted for markdown
}
