% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{open_template}
\alias{open_template}
\title{Open a New R Markdown Template}
\usage{
open_template(type = c("dummy", "report", "codebook"), ...)
}
\arguments{
\item{type}{A character vector of length 1L, representing the type of template
to be loaded. One of \emph{'dummy'}, \emph{'report'}, or
\emph{'codebook'}.}

\item{...}{Other arguments passed on to \code{\link[rmarkdown]{draft}},
apart from \code{file}, \code{template}, and \code{package}, which are
set internally.}
}
\description{
Creates an R Markdown template for various types of intermediate as well as
final project outputs such as dummy tables, analysis reports, etc.
}
