% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{show_output}
\alias{show_output}
\alias{show_output.default}
\alias{show_output.data.frame}
\title{Display output}
\usage{
show_output(x, ...)

\method{show_output}{default}(x, type = c("multiplot", "yesno"), ...)

\method{show_output}{data.frame}(
  x,
  index,
  type = c("both", "plot", "table"),
  use.regex = getOption("use.regex"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{data.frame} that has the data used in the
analysis}

\item{...}{Arguments passed on to \code{plot_multopt}.}

\item{type}{Either "table" or "plot".}

\item{index}{Integer vector representing indices for the column(s) of
interest}

\item{use.regex}{Logical.}
}
\description{
Displays some output such as a plot/
}
