% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pickout_cols_and_transform}
\alias{pickout_cols_and_transform}
\title{Pick out a Colum and Transform it}
\usage{
pickout_cols_and_transform(data, chklist)
}
\arguments{
\item{data}{A data frame}

\item{chklist}{An object of class \code{colCheck}}
}
\description{
Pick out a Colum and Transform it
}
