% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prepare_extended_col}
\alias{prepare_extended_col}
\title{Manipulate Data with Multiple Choice Questions}
\usage{
prepare_extended_col(data, variable, options, multisectoral = TRUE,
  notReferral = multisectoral)
}
\arguments{
\item{data}{A data frame with the collaped column}

\item{variable}{A character vector of length 1 representing the
name of the column to ve extended}

\item{options}{The would-be levels of the resulting factor column}

\item{multisectoral}{Logical. Whether work is on combined data frame}

\item{notReferral}{Logical. Whether not Referral Directory data}
}
\value{
A modified data frame with the choices now appearing as
levels of a factor.
}
\description{
Manipulate Data with Multiple Choice Questions
}
