% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.Tools}
\alias{print.Tools}
\title{Print Out The Tools Object}
\usage{
\method{print}{Tools}(x)
}
\arguments{
\item{x}{A \code{Tools} object}
}
\description{
Print the \code{tool.sectors} object in an easy to read format.
}
