% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{save_as_rds}
\alias{save_as_rds}
\title{Save an object with specific details}
\usage{
save_as_rds(dir, obj, filelabel, ..., state)
}
\arguments{
\item{dir}{The directory to which to save the RDS file}

\item{obj}{The object to be save.}

\item{filelabel}{A label to be applied to the filename.}

\item{...}{Additional strings to add to filename}

\item{state}{The State for whose data we are saving (specific to
RAAMP-GBV project)}
}
\description{
This is a custom wrapper for \code{saveRDS}.
}
