% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_multiopt}
\alias{table_multiopt}
\alias{table_singleopt}
\alias{table_yesno}
\title{Make a frequency tabulation}
\usage{
table_multiopt(
  data,
  dictionary = NULL,
  indices,
  use.regex = getOption("use.regex"),
  data.only = FALSE,
  redcap = getOption("data.on.redcap"),
  ...
)

table_singleopt(data, x, y, table.only = FALSE, data.only = FALSE)

table_yesno(data, col, data.only = FALSE, ...)
}
\arguments{
\item{data}{The data}

\item{dictionary}{The data dictionary}

\item{indices}{Numeric vector representing indices of columns chosen}

\item{use.regex}{Logical; whether to use label patterns to determine
what the options should be for the responses}

\item{data.only}{Return only the \code{data.frame}?}

\item{redcap}{Logical. Is the project where the function is being used one
for which the data were stored on REDCap?}

\item{...}{Additional response options, if they exist.}

\item{x, y}{An integer or character vector of length \code{1L} for selecting
a column from \code{data}.}

\item{table.only}{Logical}

\item{col}{The column with the responses}
}
\value{
An object of class \code{flextable} (or when \code{table.only} is
\code{TRUE} an object of class \code{table}).
}
\description{
Make a frequency tabulation

Make a frequency tabulation that for variables with Yes/No responses
}
\note{
To control what happens via the \code{use.regex} argument, the
project has to set this option, ideally in the local \code{.Rprofile} to
a desired value. For instance, to set a project-wide default value of
\code{FALSE} for this argument, set \code{options(use.regex = FALSE)}, and
that way, there is no need to set the argument, each time the function
is called.
}
