% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_multiopt}
\alias{table_multiopt}
\alias{table_yesno}
\title{Make a frequency tabulation of multi-response questions}
\usage{
table_multiopt(
  x,
  dictionary = NULL,
  indices,
  use.regex = getOption("use.regex"),
  data.only = FALSE,
  ...
)

table_yesno(data, col, data.only = FALSE, ...)
}
\arguments{
\item{x}{The data}

\item{dictionary}{The data dictionary}

\item{indices}{Numeric vector representing indices of columns chosen}

\item{use.regex}{Logical; whether to use label patterns to determine
what the options should be for the responses}

\item{data.only}{Return only the \code{data.frame}?}

\item{...}{Additional response options, if they exist.}

\item{data}{The data.}

\item{col}{The column with the responses}
}
\description{
Make a frequency tabulation of multi-response questions

Make a frequency tabulation that for variables with Yes/No responses
}
